+++ src/DynamoCore/Extensions/ExtensionLibraryLoader.cs
@@ -1,0 +1,25 @@
+﻿using Dynamo.Interfaces;
+using Dynamo.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using System.Text;
+
+namespace Dynamo.Extensions
+{
+    public class ExtensionLibraryLoader : ILibraryLoader
+    {
+        private readonly DynamoModel model;
+
+        internal ExtensionLibraryLoader(DynamoModel model)
+        {
+            this.model = model;
+        }
+
+        public void LoadNodeLibrary(Assembly library)
+        {
+            model.LoadNodeLibrary(library);
+        }
+    }
+}
