+++ src/DynamoCore/Interfaces/ICustomNodeManager.cs
@@ -8,6 +8,6 @@ namespace Dynamo.Interfaces
     {
         IEnumerable<CustomNodeInfo> AddUninitializedCustomNodesInPath(string path, bool isTestMode, bool isPackageMember = false);
         Guid GuidFromPath(string path);
-        bool TryGetFunctionWorkspace(Guid id, bool isTestMode, out CustomNodeWorkspaceModel ws);
+        bool TryGetFunctionWorkspace(Guid id, bool isTestMode, out ICustomNodeWorkspaceModel ws);
     }
 }
