+++ src/DynamoCore/Interfaces/IWorkspaceModel.cs
@@ -7,7 +7,7 @@ namespace Dynamo.Interfaces
     public interface IWorkspaceModel
     {
         string Name { get; }
-        string FileName { get; }
+        string FileName { get; set; }
 
         IEnumerable<ConnectorModel> Connectors { get; }
         IEnumerable<NodeModel> Nodes { get; }
