+++ src/DynamoPackages/Interfaces/CustomNodePathRemapper.cs
@@ -37,7 +37,7 @@ namespace Dynamo.PackageManager
         {
             var id = customNodeManager.GuidFromPath(originalPath);
 
-            CustomNodeWorkspaceModel def;
+            ICustomNodeWorkspaceModel def;
             var res = customNodeManager.TryGetFunctionWorkspace(id, this.isTestMode, out def);
 
             if (!res) return false;
